app/requirements.txt
paddleocr>=2.0.1
flask
gunicorn
flask-sqlalchemy
flask-sqlalchemy-session
celery
openai